+++ Sources/INCommons/UIKit/UIImage+UIView.swift
@@ -6,9 +6,10 @@ public extension UIImage {
 
 	 - parameter view: The view to create am image from.
 	 */
+	@MainActor
 	convenience init(view: UIView) {
 		UIGraphicsBeginImageContextWithOptions(view.bounds.size, view.isOpaque, 0.0)
-		view.drawHierarchy(in: view.bounds, afterScreenUpdates: false)
+		_ = view.drawHierarchy(in: view.bounds, afterScreenUpdates: false)
 		let image = UIGraphicsGetImageFromCurrentImageContext()
 		UIGraphicsEndImageContext()
 		self.init(cgImage: (image?.cgImage)!)