+++ Scripts/generateDocs.sh
@@ -2,5 +2,7 @@
 # Generate the lib's documentation.
 # Intended to be run manually from within the project folder.
 
+# Update gems.
+bundle update
 # Generate the documentation.
 bundle exec jazzy