+++ Sources/INCommonsExample/HomeView.swift
@@ -11,8 +11,10 @@ struct FunctionalityRow: View {
 struct HomeView: View {
 	let exampleViews = [
 		ExampleViewFactory(title: "Configuration", view: { AnyView(ConfigurationView()) }),
+		ExampleViewFactory(title: "Clear Button", view: { AnyView(ClearButtonExample()) }),
 		ExampleViewFactory(title: "Device Provider", view: { AnyView(DeviceProviderView()) }),
 		ExampleViewFactory(title: "First Appear View", view: { AnyView(FirstAppearView()) }),
+		ExampleViewFactory(title: "Overlay Button", view: { AnyView(OverlayButtonExample()) }),
 		ExampleViewFactory(title: "Share Sheet", view: { AnyView(ShareSheet()) }),
 		ExampleViewFactory(title: "View Condition", view: { AnyView(ViewConditions()) }),
 		ExampleViewFactory(title: "View Frame", view: { AnyView(ViewFrame()) })