+++ Changelog.md
@@ -14,6 +14,8 @@
 - Added `NotificationCenter.ObserverBag` to have a container for notification observer tokens which get automatically removed from the notification center when released.
 - Added `UIApplication.activeWindowScene` and `UIApplication.keyWindow` to retrieve the current window.
 - Added `iso8601` as an extension to Calendar to retrieve ISO8601 calendar format.
+- Fixed deprecation warning for `UIColor+KeyedArchiver`.
+- Added conformance for `Codable` and `Equatable` to `UIColor.RgbaColor`.
 
 ### v3.1.0